requirements.txt
# Core Flask packages
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

# Database
psycopg2-binary==2.9.7

# AWS S3 Integration
boto3==1.34.69
botocore==1.34.69

# File type detection for upload validation
python-magic==0.4.27

# Environment variables
python-dotenv==1.0.0

# API validation and serialization (REQUIRED)
marshmallow==4.0.0

# JWT authentication (if you're using JWT tokens)
PyJWT==2.8.0

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0

# Production server
gunicorn==21.2.0

# Security
cryptography==41.0.7
bcrypt==4.0.1

# File upload enhancements
requests==2.31.0
validators==0.22.0

# Background task processing (if using Celery)
celery==5.3.4
redis==5.0.1